+++ src/modules/access-wallet/ModuleAccessWalletMobile.vue
@@ -60,6 +60,7 @@ import {
 import { ROUTES_WALLET } from '@/core/configs/configRoutes';
 
 import handlerAnalytics from '@/modules/analytics-opt-in/handlers/handlerAnalytics.mixin';
+import { ACCESS_WALLET } from '@/modules/analytics-opt-in/handlers/configs/events';
 
 export default {
   name: 'ModuleAccessWalletMobile',
@@ -81,7 +82,9 @@ export default {
           label: 'WalletConnect',
           icon: require('@/assets/images/icons/icon-wallet-connect.svg'),
           fn: () => {
-            this.trackAccessWalletAmplitude('MobileAppWalletConnect');
+            this.trackAccessWalletAmplitude(
+              ACCESS_WALLET.MOBILE_APP_WALLET_CONNECT
+            );
             this.openWalletConnect();
           }
         },
@@ -89,7 +92,9 @@ export default {
           label: 'WalletLink',
           icon: require('@/assets/images/icons/icon-wallet-link.png'),
           fn: () => {
-            this.trackAccessWalletAmplitude('MobileAppWalletLink');
+            this.trackAccessWalletAmplitude(
+              ACCESS_WALLET.MOBILE_APP_WALLET_LINK
+            );
             this.openWalletLink();
           }
         }
@@ -100,16 +105,20 @@ export default {
     ...mapActions('wallet', ['setWallet']),
     openWalletConnect() {
       try {
-        this.trackAccessWalletAmplitude('WalletConnectQRShown');
+        this.trackAccessWalletAmplitude(ACCESS_WALLET.WALLET_CONNECT_QR_SHOWN);
         WalletConnectWallet()
           .then(_newWallet => {
             this.setWallet([_newWallet]).then(() => {
-              this.trackAccessWalletAmplitude('WalletConnectQRSuccessful');
+              this.trackAccessWalletAmplitude(
+                ACCESS_WALLET.WALLET_CONNECT_QR_SUCCESSFUL
+              );
               this.$router.push({ name: ROUTES_WALLET.DASHBOARD.NAME });
             });
           })
           .catch(e => {
-            this.trackAccessWalletAmplitude('WalletConnectQRFailed');
+            this.trackAccessWalletAmplitude(
+              ACCESS_WALLET.WALLET_CONNECT_QR_FAILED
+            );
             WalletConnectWallet.errorHandler(e);
           });
       } catch (e) {
@@ -118,16 +127,20 @@ export default {
     },
     openWalletLink() {
       try {
-        this.trackAccessWalletAmplitude('WalletLinkQRShown');
+        this.trackAccessWalletAmplitude(ACCESS_WALLET.WALLET_LINK_QR_SHOWN);
         WalletLinkWallet()
           .then(_newWallet => {
             this.setWallet([_newWallet]).then(() => {
-              this.trackAccessWalletAmplitude('WalletLinkQRSuccessful');
+              this.trackAccessWalletAmplitude(
+                ACCESS_WALLET.WALLET_LINK_QR_SUCCESSFUL
+              );
               this.$router.push({ name: ROUTES_WALLET.DASHBOARD.NAME });
             });
           })
           .catch(e => {
-            this.trackAccessWalletAmplitude('WalletLinkQRFailed');
+            this.trackAccessWalletAmplitude(
+              ACCESS_WALLET.WALLET_LINK_QR_FAILED
+            );
             WalletLinkWallet.errorHandler(e);
           });
       } catch (e) {