+++ src/modules/analytics-opt-in/handlers/configs/swap.js
@@ -1,0 +1,15 @@
+export default {
+  BROADCASTED: 'Broadcasted',
+  RECEIPT: 'Receipt',
+  FAILED: 'Failed',
+  NOT_BROADCASTED: 'NotBroadcasted',
+  VERIFY_PAGE_SHOWN: 'VerifyPageShown',
+  CANCELLED: 'Cancelled',
+  SUCCESS: 'Success',
+  CONFIRM_CLICKED: 'ConfirmClicked',
+  FIELD_INPUTS: 'FieldInputs',
+  NEXT_CLICKED: 'NextClicked',
+  INITIAL_MODAL_CLOSED: 'InitialModalClosed',
+  SWAP_INITIATED: 'SwapInitiated',
+  TRANSACTION_INITIATIED: 'TransactionInitiated'
+};