+++ src/modules/buy-sell/components/BuyComponent.vue
@@ -137,12 +137,12 @@ import { getCurrency } from '@/modules/settings/components/currencyList';
 import { buyContracts } from './tokenList';
 import { MAIN_TOKEN_ADDRESS } from '@/core/helpers/common';
 import { ETH, BSC, MATIC } from '@/utils/networks/types';
+import ModuleAddressBook from '@/modules/address-book/ModuleAddressBook.vue';
 
 export default {
   name: 'ModuleBuyEth',
   components: {
-    ModuleAddressBook: () =>
-      import('@/modules/address-book/ModuleAddressBook.vue'),
+    ModuleAddressBook: ModuleAddressBook,
     BuySellTokenSelect: () =>
       import('@/modules/buy-sell/components/TokenSelect.vue')
   },
@@ -250,7 +250,7 @@ export default {
       return fromWei(BigNumber(this.gasPrice).times(21000).toString());
     },
     priceOb() {
-      return !isEmpty(this.fetchedData)
+      return !isEmpty(this.fetchedData) && this.fetchedData[0].prices.length > 0
         ? this.fetchedData[0].prices.find(
             item => item.fiat_currency === this.selectedFiatName
           )
@@ -379,9 +379,10 @@ export default {
       return formatFloatingPointValue(this.simplexQuote.crypto_amount).value;
     },
     fiatCurrencyItems() {
-      const arrItems = this.hasData
-        ? this.fetchedData[0].fiat_currencies.filter(item => item !== 'RUB')
-        : ['USD'];
+      const arrItems =
+        this.hasData && this.fetchedData[0].fiat_currencies.length > 0
+          ? this.fetchedData[0].fiat_currencies.filter(item => item !== 'RUB')
+          : ['USD'];
       return getCurrency(arrItems);
     },
     max() {
@@ -610,7 +611,10 @@ export default {
           this.compareQuotes();
         })
         .catch(e => {
-          Toast(e, {}, ERROR);
+          const error = e.response ? e.response.data.error : e;
+          this.loading = false;
+          this.$emit('simplexQuote', {});
+          Toast(error, {}, ERROR);
         });
     },
     compareQuotes() {